refactor(core,eval): address age/ricotta sweep findings
- safety: wrap DaemonBackend socket/JSON errors in clear RuntimeError
  rather than propagating cryptic socket.timeout/OSError mid-eval
- safety: reject --json combined with --cross-validate in eval_cli
  (cross-validation path doesn't accumulate case_results, producing
  silent empty JSONL output)
- arch: extract Example, render_prompt, _parse_examples, _format_examples
  from core/rules.py into core/examples.py (rules.py: 346 → 285 lines)
- encap: route eval_cli imports through .core and .server crusts
  instead of reaching into .core.rules and .server.daemon_backend
- yagni: delete duplicate _find_project_root (use core.find_project_root)
  and dead --backend CLI arg with a single fixed choice
- ricotta: remove docstrings that restate the function name
  (_build_parser, _emit_jsonl, _format_examples, _truncate_for_event)
